--- a/.flowconfig
+++ b/.flowconfig
@@ -4,6 +4,7 @@
 .*node_modules/grpc/*
 .*node_modules/tayden-clusterfck
 .*history/*
+.*node_modules/bcryptjs
 
 [include]