--- a/e2e/scripts/runTests.babel.js
+++ b/e2e/scripts/runTests.babel.js
@@ -1,3 +1,2 @@
-require('babel-register');
-require('babel-polyfill');
-require('./runTests');
+require('babel-register')
+require('./runTests')