--- a/src/client/client.js
+++ b/src/client/client.js
@@ -6,29 +6,11 @@ import './scrollIntoViewIfNeeded'
 
 const initialState = {}
 const store = makeStore(initialState)
-
 render(<Root store={store} />, document.getElementById('root'))
 
-/*
- * react-hot-loader currently doesn't support React 16.3, so we'll need to disable it for now
- * We'll revisit this when we up our babel game #1762
- */
-
-// eslint-disable-next-line global-require
-// const {AppContainer} = require('react-hot-loader'); // eslint-disable-line import/no-extraneous-dependencies
-// render(
-// {/*<Root store={store} />,*/}
-// document.getElementById('root')
-// );
-
-// if (module.hot) {
-// module.hot.accept('./Root', () => {
-// const Root = require('./Root').default
-// render(
-// <AppContainer>
-// <Root store={store} />
-// </AppContainer>,
-// document.getElementById('root')
-// )
-// })
-// }
+if (module.hot) {
+  module.hot.accept('./Root', () => {
+    const Root = require('./Root').default
+    render(<Root store={store} />, document.getElementById('root'))
+  })
+}