--- a/src/client/client.js
+++ b/src/client/client.js
@@ -3,6 +3,8 @@ import {render} from 'react-dom'
 import makeStore from './makeStore'
 import Root from './Root'
 import './scrollIntoViewIfNeeded'
+// do this here so useBuiltIns can replace it with only the polyfills required to hit browser targets
+import 'babel-polyfill'
 
 const initialState = {}
 const store = makeStore(initialState)