--- a/src/server/createSSR.js
+++ b/src/server/createSSR.js
@@ -34,7 +34,10 @@ const extraHead = `
  <link rel="stylesheet" type="text/css" href=${fontAwesomeUrl} />
  <script>${dehydrate('__ACTION__', clientIds)}</script>
  `
-const extraBody = `<script>${segmentSnippet}</script>`
+const devBody = prod
+  ? ''
+  : '<script src="/static/vendors.dll.js"></script><script src="/static/app.js"></script>'
+const extraBody = `<script>${segmentSnippet}</script>${devBody}`
 
 const finalHTML = html
   .replace('</head>', `${extraHead}</head>`)