--- a/src/server/database/cloneProdToDev.babel.js
+++ b/src/server/database/cloneProdToDev.babel.js
@@ -1,3 +1,2 @@
 require('babel-register') // eslint-disable-line
-require('babel-polyfill')
 require('./cloneProdToDev').default()