--- a/src/server/database/migrate.babel.js
+++ b/src/server/database/migrate.babel.js
@@ -1,5 +1,4 @@
 require('babel-core/register')
-require('babel-polyfill')
 const path = require('path')
 const resolve = require('resolve') // eslint-disable-line import/no-extraneous-dependencies