--- a/src/server/utils/postDeploy.babel.js
+++ b/src/server/utils/postDeploy.babel.js
@@ -1,6 +1,5 @@
 const path = require('path')
 const resolve = require('resolve') // eslint-disable-line import/no-extraneous-dependencies
-require('babel-polyfill')
 require('babel-register')({
   // eslint-disable-line import/no-extraneous-dependencies
   resolveModuleSource (source, filename) {