--- a/src/universal/components/DashboardWrapper/DashboardWrapper.js
+++ b/src/universal/components/DashboardWrapper/DashboardWrapper.js
@@ -38,9 +38,9 @@ const query = graphql`
 `
 
 const userDashboard = () =>
-  System.import('universal/modules/userDashboard/components/UserDashboard/UserDashboard')
-const teamRoot = () => System.import('universal/modules/teamDashboard/components/TeamRoot')
-const newTeam = () => System.import('universal/modules/newTeam/containers/NewTeamForm/NewTeamRoot')
+  import('universal/modules/userDashboard/components/UserDashboard/UserDashboard')
+const teamRoot = () => import('universal/modules/teamDashboard/components/TeamRoot')
+const newTeam = () => import('universal/modules/newTeam/containers/NewTeamForm/NewTeamRoot')
 
 const subscriptions = [
   NewAuthTokenSubscription,