--- a/src/universal/components/DashboardWrapper/DashboardWrapper.js
+++ b/src/universal/components/DashboardWrapper/DashboardWrapper.js
@@ -38,9 +38,11 @@ const query = graphql`
 `
 
 const userDashboard = () =>
-  import('universal/modules/userDashboard/components/UserDashboard/UserDashboard')
-const teamRoot = () => import('universal/modules/teamDashboard/components/TeamRoot')
-const newTeam = () => import('universal/modules/newTeam/containers/NewTeamForm/NewTeamRoot')
+  import(/* webpackChunkName: 'UserDashboard' */ 'universal/modules/userDashboard/components/UserDashboard/UserDashboard')
+const teamRoot = () =>
+  import(/* webpackChunkName: 'TeamRoot' */ 'universal/modules/teamDashboard/components/TeamRoot')
+const newTeam = () =>
+  import(/* webpackChunkName: 'NewTeamRoot' */ 'universal/modules/newTeam/containers/NewTeamForm/NewTeamRoot')
 
 const subscriptions = [
   NewAuthTokenSubscription,