--- a/src/universal/components/QueryRenderer/QueryRenderer.js
+++ b/src/universal/components/QueryRenderer/QueryRenderer.js
@@ -327,4 +327,4 @@ function fetchQueryAndComputeStateFromProps(
 }
 
 polyfill(ReactRelayQueryRenderer)
-module.exports = ReactRelayQueryRenderer
+export default ReactRelayQueryRenderer