--- a/src/universal/components/SuggestMentionableUsersRoot.js
+++ b/src/universal/components/SuggestMentionableUsersRoot.js
@@ -26,10 +26,8 @@ const loading = (props) => (
 )
 const LoadableSuggestMentionableUsers = Loadable({
   loader: () =>
-    System.import(
-      /* webpackChunkName: 'SuggestMentionableUsers' */
-      'universal/components/SuggestMentionableUsers'
-    ),
+    import(/* webpackChunkName: 'SuggestMentionableUsers' */
+      'universal/components/SuggestMentionableUsers'),
   loading,
   delay: HUMAN_ADDICTION_THRESH,
   timeout: MAX_WAIT_TIME