This is about a year overdue, and it feels so, so good. this is huge for any user that has less than wonderful internet.

Changes:
- move to webpack 4. kinda painful, but the new chunking strategy is really smart!
- removed SSR. now we just send a skeleton & fill it in with JS (fix #2159)
- updated our babel strategy. we were using things that have been deprecated for over a year (fix #1762)
- clean up scripts (fix #1407)
- remove babel-polyfill. all the browsers we want to support (including mobile!) have supported await/async for a good long while now: https://caniuse.com/#feat=async-functions
- move to bcryptjs from bcrypt. faster builds, now folks on windows can build. C++ to .node was a pain.
- name all our webpack chunks for easy debugging
- remove build:theme. we originally did this to keep tinycolor2 out of our build, but that thing is used all over the place, so we might as well exclude the appTheme.json & just build it on the client. smaller payload!
- async load signin/signup page. auth0-js was HUGE. over 2x the size of react. now that doesn't load unless they need it
- removed class decorators (eg `@withHotkey`). TC39 is going in a different direction & it looks like that won't be valid javascript for awhile
- removed support for IE11. we had to do some crazy transpiling to support it, which affects speed & initial payload size.

DX Results:
- Build time down from ~6 seconds to ~3 seconds (cached)
- CI Build time down ~5 mins
- add live reloading (real hot-reloading without refresh is stuck on a bug. react-hot-loader's babel plugin doesn't play nicely with relay) 

UX Results:
- __no flashing the signin page__
- __Initial payload size down 50% (1.4MB to 740KB)__
- __DOM Content load time down 40% (575ms to 300ms)__
- __Site load time down 20% (1000ms to 800ms)__

TEST
- standard integration test will find any problems