@jordanh 2018-06-05
# Test Cycle 1
# Release Integration Test Checklist
**Status:** (in progress/completed)
​
## Tests:
​
### Authentication
- [ ] Can sign up
- [ ] Can sign in
- [ ] Can sign out
- [ ] Sees warning when login attempt with wrong creds
​
### Team Dashboard
- [ ] Can navigate to Team Dashboard
- [ ] Can create a Task Card on Team Dashboard
- [ ] Can create 3 additional Task Cards on Team Dashboard
- [ ] Can re-assign ownership of Task Card
- [ ] Can resort priority among 3 Task Cards
- [ ] Can assign new Task status via click
- [ ] Can assign new Task status via drag
- [ ] Can navigate to second team from Team Dashboard
- [ ] Can create Task Card on second team
- [ ] Can create 3 Agenda Topics on Team Dashboard
- [ ] Can reorder Agenda Topics on Team Dashboard
- [ ] Can invite a new teammate from the Assign Owner menu on a Task Card
​
### Task Card Content
- [ ] Can add, edit, navigate to, and remove a link in a Task card
- [ ] Can format text (bold, italics, underline, strikethrough [press ? for formatting options])
- [ ] Can add emoji
- [ ] Can mention a teammate
- [ ] Can add #private tag by typing/using inline menu
- [ ] Can add #archived tag by typing/using inline menu
- [ ] Can add code, codeblock, quote content
​
### Team Dashboard > Archive
- [ ] Can archive 1 Task Card
- [ ] Can navigate to Archive
- [ ] Can restore Task Card
​
### Team Dashboard > Settings
- [ ] Can navigate to Team Settings
- [ ] Can invite new user by email address
- [ ] Can add Slack integration and channel
- [ ] Can add GitHub integration and repository
​
### Accept Email Invitation
- [ ] New user can accept email invitation
​
### Action Meeting
- [ ] Can navigate to Meeting Lobby
- [ ] Can see presence of second Team Member
- [ ] Can start Meeting from Meeting Lobby
- [ ] Can observe Slack meeting start notification on channel
- [ ] Facilitator can lead meeting through Check-In
- [ ] Can create Task Cards on Updates dashboard
- [ ] Participant can follow, diverge and rejoin Facilitator
- [ ] Participant can add Agenda Topic
- [ ] Can reorder Agenda Topics
- [ ] Facilitator can proceed to process Agenda Topics
- [ ] Participant can add Task Card on Agenda Topic
- [ ] Participant can add second Task Card on Agenda Topic
- [ ] Participant can promote second Task Card to GitHub issue
- [ ] Participant can assign a new Task Card to another teammate
- [ ] Can invite a new teammate from the Assign Owner menu on a Task Card
- [ ] Facilitator can proceed through Agenda Topics to Last Call
- [ ] Facilitator can end Meeting
- [ ] All Participants can see Meeting Summary
- [ ] All Participants receive email Meeting Summary
- [ ] Can observe Slack meeting stop notification on channel
​
### Retro Meeting
- [ ] 1 user can start the retro
- [ ] The facilitator can complete the Check-In phase
- [ ] 2 users can add reflection cards
- [ ] 2 users can group
- [ ] 2 users can vote
- [ ] 2 users can add tasks during the Discuss phase
- [ ] The facilitator can end the meeting
- [ ] Users can see summary in browser
- [ ] Users receive summary email
​
### Notifications
- [ ] Can assign new other user, other user can see notification
- [ ] Can @mention card to other user, other user can see noficiation
- [ ] User can clear notifications
​
### My Dashboard
- [ ] Can navigate to My Dashboard
- [ ] Can reassign priority across teams from My Dashboard
​
### Team Dashboard > Settings > Manage Team
- [ ] Can remove Team Member
​