@jordanh 2018-06-05
Test cycle complete. I don't think the invite thing is a regression. Merge please review results and merge if you wish...